atualização do adicionar ao cookie/bd a ficha alimentar

+++ src/main/java/com/CaloriesCalculator/dto/ProdutoAlimenticioDto.java
@@ -3,14 +3,12 @@ package com.CaloriesCalculator.dto;
 // representa cada alimento individual dentro da lista
 public class ProdutoAlimenticioDto {
 
-    private Integer id;
+    private Long id;
     private String name;
     private CategoryDTO category;
     private NutrientsDTO nutrients;
 
-    public Integer getId() {
-        return id;
-    }
+    public Long getId() {return id;}
 
     public String getName() {
         return name;
